chore: moving away files operations to separate filesystem package
